Actual fix for problems with buffer overruns
It appears that Go strings are not always null-terminated, and the
C-for-Go helper functions simply translated Go strings into C
pointers. The C API side then proceeded to read heaps of data it
wasn't supposed to read, into a Vorbis comment, and produce invalid
packets.

In the past, we've attempted to fix it by disabling compiler
optimizations, as in some circumstances it appeared to have fixed
the issue. However, since we've now discovered the actual culprit,
we add null-terminators to comment strings, and re-enable all
optimizations.

Signed-off-by: Burillo <11786086+burillo-se@users.noreply.github.com>